correction

+++ src/search.h
@@ -70,7 +70,7 @@ struct RootMove {
   Value previousScore = -VALUE_INFINITE;
   int selDepth = 0;
   int tbRank = 0;
-  int nodesSearched = 0;
+  uint64_t nodesSearched = 0;
   Value tbScore;
   std::vector<Move> pv;
 };
